Move message box tests to display tests.
